+++ jy-admin-dal/src/main/java/org/loxf/jyadmin/dal/po/AccountDetail.java
@@ -2,6 +2,7 @@ package org.loxf.jyadmin.dal.po;
 
 import org.loxf.jyadmin.base.bean.BaseModel;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 public class AccountDetail extends BaseModel {
@@ -13,9 +14,9 @@ public class AccountDetail extends BaseModel {
 
     private Integer type;
 
-    private Long balance;
+    private BigDecimal balance;
 
-    private Long changeBalance;
+    private BigDecimal changeBalance;
 
     public String getCustId() {
         return custId;
@@ -49,19 +50,19 @@ public class AccountDetail extends BaseModel {
         this.type = type;
     }
 
-    public Long getBalance() {
+    public BigDecimal getBalance() {
         return balance;
     }
 
-    public void setBalance(Long balance) {
+    public void setBalance(BigDecimal balance) {
         this.balance = balance;
     }
 
-    public Long getChangeBalance() {
+    public BigDecimal getChangeBalance() {
         return changeBalance;
     }
 
-    public void setChangeBalance(Long changeBalance) {
+    public void setChangeBalance(BigDecimal changeBalance) {
         this.changeBalance = changeBalance;
     }
